# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Chat API Documentation",
        "description": "This is the API documentation for the Chat application",
        "contact": {
            "email": "support@yourapp.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/api/chats": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "summary": "Fetch all chats for the authenticated user",
                "operationId": "1b4c79f8be9c8d8682ddd42225df131b",
                "responses": {
                    "200": {
                        "description": "A list of chats",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "oneOf": [
                                            {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Group Chat"
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Private Chat"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        ]
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Chats"
                ],
                "summary": "Create a new private chat",
                "operationId": "b49d6a86afd5c3ed67550138804d654b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recipient_id"
                                ],
                                "properties": {
                                    "recipient_id": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chat created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "recipient_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Chat name"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/chats/{chat_id}": {
            "get": {
                "tags": [
                    "Chats"
                ],
                "summary": "Get chat details",
                "description": "Retrieve details of a specific chat including recipient ID and chat metadata.",
                "operationId": "b770c03485594eb8af8d5ceb4218ae8d",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "chat_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "recipient_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Chat with John Doe"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Chats"
                ],
                "summary": "Delete a chat",
                "operationId": "19d6a797637912b3e86db2b998584a8e",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Chat deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Chat deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Chats"
                ],
                "summary": "Update chat details",
                "operationId": "eb91806a5aa884517cf51ff735aa6346",
                "parameters": [
                    {
                        "name": "chat_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255,
                                        "example": "Updated Chat Name"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chat updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Updated Chat Name"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chat not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
    },
...
}